+++ tests/tests_tiles.py
@@ -29,6 +29,10 @@ class TestTiles(unittest.TestCase):
         self.assertEqual(tile.letter, "O")
         self.assertEqual(tile.value, 1)
 
+    def test_repr(self):
+        tile= Tile('A', 1)
+        self.assertEqual(repr(tile), 'A:1')
+
 class TestBagTiles(unittest.TestCase):
     @patch('random.shuffle')
     def test_bag_tiles(self, patch_shuffle):